+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Native MonitorFlare webhook adapter for `monitor.down` and `monitor.up` events
+
 ## [0.1.0] - 2026-09-17
 
 ### Added